@@ -4,7 +4,7 @@ version: 0.3.1
 homepage: https://polkawallet.io
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.22.0"
 
 dependencies:
@@ -16,17 +16,18 @@ dependencies:
   url_launcher: ^6.0.4
   intl: ^0.17.0
   qr_flutter:
-    git:
-      url: https://github.com/polkawallet-io/qr.flutter.git
-      ref: master
+    path: ../qr_flutter
+    # git:
+    #   url: https://github.com/polkawallet-io/qr.flutter.git
+    #   ref: master
   flutter_qr_scan:
-#    path: ../flutter_qr_reader
-    git:
-      url: https://github.com/polkawallet-io/flutter_qr_reader.git
-      ref: a125851ca20038767ac4efa2e20d12116a390414
-  polkawallet_sdk: ^0.3.1
-#  polkawallet_sdk:
-#    path: ../sdk
+    path: ../flutter_qr_reader
+    # git:
+    #   url: https://github.com/polkawallet-io/flutter_qr_reader.git
+    #   ref: a125851ca20038767ac4efa2e20d12116a390414
+  # polkawallet_sdk: ^0.3.1
+  polkawallet_sdk:
+   path: ../sdk
 
 dev_dependencies:
   flutter_test:
