@@ -10,16 +10,16 @@ class AccountSelectList extends StatelessWidget {
   AccountSelectList(this.plugin, this.list);
 
   final PolkawalletPlugin plugin;
-  final List<KeyPairData> list;
+  final List<KeyPairData>? list;
 
   @override
   Widget build(BuildContext context) {
     return ListView(
-      children: list.map((i) {
+      children: list!.map((i) {
         return ListTile(
           leading: AddressIcon(i.address, svg: i.icon),
           title: Text(UI.accountName(context, i)),
-          subtitle: Text(Fmt.address(i.address)),
+          subtitle: Text(Fmt.address(i.address)!),
           trailing: Icon(Icons.arrow_forward_ios, size: 16),
           onTap: () => Navigator.of(context).pop(i),
         );
